requirements.txt
asgiref==3.3.1
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.5.5
bcrypt==3.1.7
cffi==1.13.0
colorama==0.4.1
cryptography==2.8
Django==3.1.7
djangorestframework==3.12.2
Faker==3.0.0
gitdb2==2.0.6
GitPython==3.0.3
grt==0.1
importlib-metadata==0.19
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==7.2.0
mygrations==0.11.0
packaging==19.1
paramiko==2.6.0
Pillow==8.1.2
pluggy==0.12.0
py==1.8.0
pycodestyle==2.6.0
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.4.2
pytest==5.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
selenium==3.141.0
six==1.12.0
smmap2==2.0.5
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
gunicorn
django-heroku
